// backend/src/careers/enums/champion-archetype.enum.ts
export enum ChampionArchetype {
  TOP_TANK = 'TOP_TANK',
  TOP_AD_BRUISER = 'TOP_AD_BRUISER',
  TOP_AP_BRUISER = 'TOP_AP_BRUISER',
  TOP_SIDE_LANE = 'TOP_SIDE_LANE',
  TOP_VALUE = 'TOP_VALUE',
  JUNGLE_ENGAGE = 'JUNGLE_ENGAGE',
  JUNGLE_SCALING = 'JUNGLE_SCALING',
  JUNGLE_EARLY_SNOWBALL = 'JUNGLE_EARLY_SNOWBALL',
  MID_ASSASSIN = 'MID_ASSASSIN',
  MID_STANDING_MAGE = 'MID_STANDING_MAGE',
  MID_TANK = 'MID_TANK',
  MID_AP_BRUISER = 'MID_AP_BRUISER',
  MID_AD_BRUISER = 'MID_AD_BRUISER',
  MID_VALUE = 'MID_VALUE',
  LANE_BULLY = 'LANE_BULLY',
  HYPER_CARRY = 'HYPER_CARRY',
  WEAKSIDE_SAFE = 'WEAKSIDE_SAFE',
  GRAB = 'GRAB',
  UTILITY = 'UTILITY',
  TANK_ENGAGE = 'TANK_ENGAGE',
}